gpu: nvgpu: perf: update type for size param
Update interfaces to use size_t to align with boardobj_construct_super()
and avoid unnecessary casts.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2011435
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/perf/vfe_equ.c
@@ -436,7 +436,7 @@ static int vfe_equ_pmudatainit_super(struct gk20a *g,
 
 static int vfe_equ_construct_super(struct gk20a *g,
 				   struct boardobj **ppboardobj,
-				   u16 size, void *pargs)
+				   size_t size, void *pargs)
 {
 	struct vfe_equ *pvfeequ;
 	struct vfe_equ *ptmpequ = (struct vfe_equ *)pargs;
@@ -492,7 +492,7 @@ static int vfe_equ_pmudatainit_compare(struct gk20a *g,
 
 static int vfe_equ_construct_compare(struct gk20a *g,
 				     struct boardobj **ppboardobj,
-				     u16 size, void *pargs)
+				     size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct vfe_equ_compare *pvfeequ;
@@ -553,7 +553,7 @@ static int vfe_equ_pmudatainit_minmax(struct gk20a *g,
 
 static int vfe_equ_construct_minmax(struct gk20a *g,
 				    struct boardobj **ppboardobj,
-				    u16 size, void *pargs)
+				    size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct vfe_equ_minmax *pvfeequ;
@@ -611,7 +611,7 @@ static int vfe_equ_pmudatainit_quadratic(struct gk20a *g,
 
 static int vfe_equ_construct_quadratic(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct vfe_equ_quadratic *pvfeequ;
@@ -669,7 +669,7 @@ static int vfe_equ_pmudatainit_scalar(struct gk20a *g,
 
 static int vfe_equ_construct_scalar(struct gk20a *g,
 				       struct boardobj **ppboardobj,
-				       u16 size, void *pargs)
+				       size_t size, void *pargs)
 {
 	struct boardobj *ptmpobj = (struct boardobj *)pargs;
 	struct vfe_equ_scalar *pvfeequ;
